Make labels consistent and name-agnostic (#28)
* Make labels consistent and name-agnostic

We already have the weird label for X25519 + ML-KEM-768, and it looks
like the conversation around names will drag on for a while longer, so
let's just make labels consistent and name-agnostic, so we can request
IANA codepoints and ship implementations without blocking on names.

X25519 + ML-KEM-768 is a rebel X-Wing fighter.

    \./
    /^\

P-256 + ML-KEM-1024 is an imperial TIE fighter.

    |-()-|

P-384 + ML-KEM-1024 is a heavy imperial Lambda shuttle.

     |
    /-\

They are all six characters, and not more annoying than the X-Wing one
that implementers have to deal with anyway.

* add hex to label

Co-authored-by: Richard Barnes <rlb@ipv.sx>

* add hex to label

Co-authored-by: Richard Barnes <rlb@ipv.sx>

* add hex to label

---------

Co-authored-by: Deirdre Connolly <durumcrustulum@gmail.com>
Co-authored-by: Richard Barnes <rlb@ipv.sx>